Remove bundling
